修改选择题分数评判功能

+++ src/backend/api/Test.go
@@ -150,13 +150,13 @@ func QueryQuesIdByTestID(db *gorm.DB, id int) []int {
 	return quesId
 }
 
-func QueryGradeByTestIdAndQuestionId(db *gorm.DB, testId int, questionId int) (int, error) {
+func QueryGradeByTestIdAndQuestionId(db *gorm.DB, testId int, questionId int) (float64, error) {
 	// 查询该题目的分数
 	var test Tests
 	if err := db.Table("tests").Where("id = ? AND question_id = ?", testId, questionId).Find(&test).Error; err != nil {
 		return 0, err
 	}
-	return int(test.Grade), nil
+	return test.Grade, nil
 }
 
 func QueryTestsByStuName(db *gorm.DB, username string) ([]Tests, error) {
